+++ src/routes/app/index.js
@@ -1,35 +1,74 @@
 import { createAppContainer } from 'react-navigation';
 import { createStackNavigator } from 'react-navigation-stack';
-import { createDrawerNavigator } from 'react-navigation-drawer';
-
+import { createBottomTabNavigator } from 'react-navigation-tabs';
+import { SafeAreaView } from 'react-native';
 import { resize } from '../../utils/styles';
 import React from 'react';
 // screens
 import Home from '../../screens/app/Home';
-import Detail from '../../screens/app/Detail'
-// const drawerStack = createDrawerNavigator(
+import Detail from '../../screens/app/Detail';
+import Favorites from '../../screens/app/Favorites'
+//custom
+import NavBarBottom from '../../components/BottomNavBar';
+import colors from '../../utils/colors';
+import fonts from '../../utils/fonts';
+const bottomTabNavigator = createBottomTabNavigator(
+  {
+    Home: {
+      screen: Home,
+    },
+    Favorites: {
+      screen: Favorites,
+    },
+  },
+  {
+    tabBarComponent: props => (
+      <SafeAreaView>
+        <NavBarBottom {...props} />
+      </SafeAreaView>
+    ),
+    tabBarPosition: 'bottom',
+    tabBarOptions: {
+      activeTintColor: colors.lightPurple,
+      inactiveTintColor: colors.darkTextLight,
+      labelStyle: {
+        fontFamily: fonts.notoBold,
+        fontSize: resize(10),
+        fontWeight: '700',
+        letterSpacing: 0.12,
+      },
+      style: {
+        backgroundColor: colors.backgroundColorApp,
+      },
+    },
+    animationEnabled: false,
+    swipeEnabled: false,
+    lazy: true,
+    initialRouteName: 'Home',
+  },
+);
+
+// const homeStackNavigator = createStackNavigator(
 //   {
 //     Home: {
-//       screen: Home
+//       screen: Home,
 //     },
+//     Detail: {
+//       screen: Detail,
+//     }
 //   },
-
 //   {
 //     initialRouteName: 'Home',
-//     contentComponent: props => <Sidebar {...props} />,
-//     drawerWidth: resize(250),
-//     drawerPosition: 'right',
 //     defaultNavigationOptions: {
-//       header: null,
+//       headerShown: false
 //     },
 //   }
-// )
-
+// );
 
 const appStackNavigator = createStackNavigator(
   {
     Home: {
-      screen: Home,
+      screen: bottomTabNavigator,
     },
     Detail: {
       screen: Detail,